added UsersList component
