+++ client/src/components/personsList/personsList.html
@@ -1,17 +1,20 @@
 <div class="col m6 l4" ng-repeat="person in $ctrl.persons">
 	<div class="card">
 	    <div class="card-image waves-effect waves-block waves-light">
-	      <img class="activator" ng-src="{{person.picture}}">
+	      <img class="activator" ng-src="{{person.avatar}}">
 	    </div>
 	    <div class="card-content">
-	      <span class="card-title activator grey-text text-darken-4">{{person.name}}<i class="material-icons right">more_vert</i></span>
+	      <span class="card-title activator grey-text text-darken-4">
+			{{person.firstName}} {{person.lastName}}
+	      	<i class="material-icons right">more_vert</i>
+	      </span>
 	      <p><a href="#">{{person.email}}</a></p>
 	    </div>
 	    <div class="card-reveal">
 	      <span class="card-title grey-text text-darken-4">About:<i class="material-icons right">close</i></span>
 	      <p>
 	      	I am {{person.age}}.<br />
-	      	{{person.about}}
+	      	{{person.aboutMe}}
 	      </p>
 	    </div>
 	</div>
