+++ client/src/components/personsList/personsListController.js
@@ -1,0 +1,10 @@
+function UsersListController($http) {
+  let self = this;
+
+  $http.get('/api/users').then(function(res) {
+    self.persons = res.data;
+  });
+}
+UsersListController.$inject = ['$http'];
+
+module.exports = UsersListController;
